+++ src/helpers/getERC721Ledger.ts
@@ -1,6 +1,6 @@
 import { SealCredERC721Ledger } from '@big-whale-labs/seal-cred-ledger-contract'
 import Ledger from 'models/Ledger'
-import getLedgerRecord from 'helpers/getLedgerRecord'
+import getERC721LedgerRecord from 'helpers/getERC721LedgerRecord'
 
 export default async function (ledger: SealCredERC721Ledger): Promise<Ledger> {
   const eventsFilter = ledger.filters.CreateDerivativeContract()
@@ -9,7 +9,7 @@ export default async function (ledger: SealCredERC721Ledger): Promise<Ledger> {
     (prev, { args: { originalContract, derivativeContract } }) => {
       return {
         ...prev,
-        [originalContract]: getLedgerRecord(
+        [originalContract]: getERC721LedgerRecord(
           originalContract,
           derivativeContract
         ),
