+++ src/helpers/getEmailLedgerRecord.ts
@@ -1,0 +1,12 @@
+import { SCERC721Derivative__factory } from '@big-whale-labs/seal-cred-ledger-contract'
+
+import defaultProvider from 'helpers/defaultProvider'
+
+export default function (derivativeContract: string) {
+  return {
+    derivativeContract: SCERC721Derivative__factory.connect(
+      derivativeContract,
+      defaultProvider
+    ),
+  }
+}
